Remove set_key_shared function
It is no longer shared any more.
